backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
numpy
pillow
scikit-learn
google-search-results
requests
google-cloud-vision
